refactor(ProductChangePriceEvent): Fix productPrice
