+++ src/main/java/kitchenpos/products/application/MenuServiceClient.java
@@ -1,4 +1,4 @@
-package kitchenpos.support.application;
+package kitchenpos.products.application;
 
 import java.math.BigDecimal;
 import java.util.UUID;
