+++ src/main/java/kitchenpos/products/application/ProductChangePriceEvent.java
@@ -1,9 +1,11 @@
 package kitchenpos.products.application;
 
+import kitchenpos.support.domain.ProductPrice;
+
 import java.util.UUID;
 
-public record ProductChangePriceEvent(UUID productId) {
-    public static ProductChangePriceEvent from(UUID productId) {
-        return new ProductChangePriceEvent(productId);
+public record ProductChangePriceEvent(UUID productId, ProductPrice price) {
+    public static ProductChangePriceEvent from(UUID productId, ProductPrice price) {
+        return new ProductChangePriceEvent(productId, price);
     }
 }
